Fix: 유저 키,몸무게 실수형으로 변경 (#68)

+++ src/main/java/com/dnd/Exercise/domain/user/dto/request/UpdateMyProfileReq.java
@@ -24,9 +24,9 @@ public class UpdateMyProfileReq {
 
     private Integer age;
 
-    private Integer height;
+    private Double height;
 
-    private Integer weight;
+    private Double weight;
 
     private SkillLevel skillLevel;
 }
